+++ core-ng/src/main/java/core/framework/internal/web/websocket/WebSocketMessageListener.java
@@ -3,6 +3,8 @@ package core.framework.internal.web.websocket;
 import core.framework.internal.log.ActionLog;
 import core.framework.internal.log.LogManager;
 import core.framework.internal.web.http.RateControl;
+import core.framework.web.exception.BadRequestException;
+import core.framework.web.exception.TooManyRequestsException;
 import io.undertow.websockets.core.AbstractReceiveListener;
 import io.undertow.websockets.core.BufferedTextMessage;
 import io.undertow.websockets.core.CloseMessage;
@@ -45,7 +47,7 @@ final class WebSocketMessageListener extends AbstractReceiveListener {
             wrapper.handler.listener.onMessage(wrapper, message);
         } catch (Throwable e) {
             logManager.logError(e);
-            WebSockets.sendClose(CloseMessage.UNEXPECTED_ERROR, e.getMessage(), channel, ChannelCallback.INSTANCE);
+            WebSockets.sendClose(closeCode(e), e.getMessage(), channel, ChannelCallback.INSTANCE);
         } finally {
             logManager.end("=== ws message handling end ===");
         }
@@ -90,6 +92,13 @@ final class WebSocketMessageListener extends AbstractReceiveListener {
         actionLog.context("client_ip", wrapper.clientIP);
         actionLog.context("listener", wrapper.handler.listener.getClass().getCanonicalName());
         actionLog.context("room", wrapper.rooms.toArray());
+    }
+
+    // as websocket does not have restful convention, here only supports general cases
+    int closeCode(Throwable e) {
+        if (e instanceof TooManyRequestsException) return WebSocketCloseCodes.TRY_AGAIN_LATER;
+        if (e instanceof BadRequestException) return WebSocketCloseCodes.POLICY_VIOLATION;
+        return WebSocketCloseCodes.INTERNAL_ERROR;
     }
 
     @Override