+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### 7.4.9 (06/04/2020 - )
 * http: for static content controllers, convert client abort exception to warning (still better use CDN/Storage solution to serve static resources)
 * log-collector: ignore sendBeacon request with empty body, (browser/privacy/plugin may impact sendBeacon's behavior)
+* ws: support limitRate for websocket message listener
 
 ### 7.4.8 (05/26/2020 - 06/02/2020)
 * monitor: fix kube monitor false alert with too many restarts (use neowu/monitor:7.4.8-b0 before releasing 7.4.8)